Абсолютно позиционирует хедер. Выводит индекса буллитов. Анимация - WIP
