+++ src/index.html
@@ -221,7 +221,7 @@
               </ul>
             </div>
           </li>
-          <!-- <li class="periods__item period swiper-slide">
+          <li class="periods__item period swiper-slide">
             <h2 class="period__current">1989&nbsp;&nbsp;<span>1993</span></h2>
             <h3 class="period__title">Nirvana</h3>
             <div class="period__swiper swiper">
@@ -302,7 +302,7 @@
                 </div>
               </ul>
             </div>
-          </li> -->
+          </li>
         </ul>
       </div>
       <div class="periods__controls controls">
@@ -343,6 +343,7 @@
         </div>
       </div>
     </section>
+    
   </main>
 
   <script src="./bundle.js"></script>
