+++ src/sass/blocks/controls/controls.scss
@@ -143,7 +143,6 @@
       &--active {
         width: 56px;
         height: 56px;
-        border-radius: 50%;
         background-color: $grey;
         border: 1px solid $dark-grey-blue;
         text-align: center;
@@ -153,22 +152,5 @@
         transform: translate(-49.5%, -49.5%);
       }
     }
-
-    // &--active {
-    //   background-color: $dark-grey-blue;
-
-    //   @media (min-width: $desktop-view) {
-    //     width: 56px;
-    //     height: 56px;
-    //     border-radius: 50%;
-    //     background-color: $grey;
-    //     border: 1px solid $dark-grey-blue;
-    //     text-align: center;
-    //     vertical-align: middle;
-    //     font-size: 20px;
-    //     line-height: 30px;
-    //     transform: translate(-49.5%, -49.5%);
-    //   }
-    // }
   }
 }
