+++ src/ts/main.ts
@@ -1,5 +1,3 @@
 import '../sass/style.scss';
 
-import './pagination';
-// import './sliders/date-slider';
-import './sliders/periods-slider';
+import './periods/index';
