+++ src/ts/periods/index.ts
@@ -1,0 +1,4 @@
+import './pagination';
+import { initPeriodsSlider } from './sliders/periods-slider';
+
+window.addEventListener('load', initPeriodsSlider);
