+++ src/ts/periods/sliders/date-slider.ts
